tools/depends.c: Pass the context pointer to tal_ functions
Use the context pointer passed into our functions as the root
to allocate new objects from. This makes it clearer that we
are passing in a correct pointer when debugging.

This came up because of duplicate symbols between modules that
are incompatible because of their choice of tal vs talloc, when
linking against a monolithic libccan.a. In this case the talloc
version of grab_file was being called with a tal pointer, and
crashing as a result when checked next by tal.

We may want to consider forcing all global symbols be be unique
between modules, so that linking with libccan.a works.

Signed-off-by: Jon Griffiths <jon_p_griffiths@yahoo.com>